Refactor out sortByDependencies
